backend/requirements.txt
flask
flask-cors
mysql.connector
python-dotenv